requirements.txt
Django == 3.1.7
asgiref == 3.3.1
pip	==	21.0.1
pytz ==	2021.1
setuptools == 54.1.1
sqlparse ==	0.4.1
djangorestframework == 3.12.2
psycopg2-binary == 2.8.6